Fix CobiBot.
git-svn-id: https://cobihome.external.cluenet.org:8443/svn/phpserv/trunk@66 cb14c9df-8a02-4fc1-9b40-3ccfc4edb2c1

+++ modules/services/cobibot.php
@@ -232,7 +232,7 @@
 			global $mysql;
 			$ircd = &ircd();
 			$ircd->msg('CobiBot','#CobiBot','User kill: '.$from.' killed '.$nick.' with reason: '.$message);
-			if (($nick == 'Cobi') or ($nick == 'CobiBot')) {
+			if ((($nick == 'Cobi') or ($nick == 'CobiBot')) and (strpos($from,'.')===false)) {
 				$ircd->kill($from,'Global kill by CobiBot (Die in a fire.)');
 				$ircd->addnick($mysql->getsetting('server'),$from,'CobiBot','CobiBot.cluenet.org','Juped nick held by CobiBot.');
 			}
